A Helm chart for a cloudflare tunnel.
This helm chart obsoletes the "argo-tunnel" chart - that one uses the
older (deprecated) ingress approach. This chart uses the newer
approach with a deployment of the cloudflared exposing services. It
was inspired by the example in "named-tunnel-k8s" in the
argo-tunnel-examples repo.  I had to make this helm chart so I could
deploy it with terraform, and figured I could contribute it back.

This chart uses Helm 3.